+++ Common/cpp/Tools/ReanimatedJSIUtils.cpp
@@ -1,4 +1,4 @@
-#include "Common/cpp/Tools/JsiUtils.h"
+#include "ReanimatedJSIUtils.h"
 #include <vector>
 
 using namespace facebook;
